+++ app/src/Controller/NamespacedMoodleNetController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Controller;
+
+use Symfony\Component\Routing\Annotation\Route;
+
+#[Route('/{serverID}/.pkg/@moodlenet', name: 'namedspaced_moodlenet_')]
+class NamespacedMoodleNetController extends MoodleNetController
+{
+}
